+++ intelmq/lib/mixins/cache.py
@@ -1,4 +1,4 @@
-""" CacheMixin for IntelMQ
+"""CacheMixin for IntelMQ
 
 SPDX-FileCopyrightText: 2021 Sebastian Waldbauer
 SPDX-License-Identifier: AGPL-3.0-or-later
@@ -6,12 +6,26 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 CacheMixin is used for caching/storing data in redis.
 """
 
+import json
 from typing import Any, Optional
 import redis
 import intelmq.lib.utils as utils
 
 
 class CacheMixin:
+    """Provides caching possibilities for bots, see also https://docs.intelmq.org/latest/dev/bot-development/#mixins
+
+    For key-value cache, use methods:
+        cache_exists
+        cache_get
+        cache_set
+
+    To store dict elements in a cache queue named after bot id, use methods:
+        cache_put
+        cache_pop
+        cache_length
+    """
+
     __redis: redis.Redis = None
     redis_cache_host: str = "127.0.0.1"
     redis_cache_port: int = 6379
@@ -31,7 +45,9 @@ class CacheMixin:
                 "socket_timeout": 5,
             }
 
-        self.__redis = redis.Redis(db=self.redis_cache_db, password=self.redis_cache_password, **kwargs)
+        self.__redis = redis.Redis(
+            db=self.redis_cache_db, password=self.redis_cache_password, **kwargs
+        )
         super().__init__()
 
     def cache_exists(self, key: str):
@@ -50,6 +66,20 @@ class CacheMixin:
         self.__redis.set(key, value)
         if self.redis_cache_ttl:
             self.__redis.expire(key, self.redis_cache_ttl)
+
+    def cache_put(self, value: dict) -> int:
+        # Returns the length of the list after pushing
+        size = self.__redis.lpush(self.bot_id, json.dumps(value))
+        return size
+
+    def cache_length(self) -> int:
+        return self.__redis.llen(self.bot_id)
+
+    def cache_pop(self) -> dict:
+        data = self.__redis.rpop(self.bot_id)
+        if data is None:
+            return None
+        return json.loads(data)
 
     def cache_flush(self):
         """